requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==43.0.3
alembic==1.13.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
aiofiles==24.1.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
Pillow==11.0.0
boto3==1.35.36
supabase==2.9.1